% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{cxxflags}
\alias{cxxflags}
\title{cxxflags}
\usage{
cxxflags(as_character = FALSE)
}
\arguments{
\item{as_character}{Whether the result should be returned as a string}
}
\description{
Function for returning the C/C++ flags needed for compilation using the package's headers
}
