% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qjs.R
\name{from_json}
\alias{from_json}
\title{from_json}
\usage{
from_json(json)
}
\arguments{
\item{json}{JSON string to convert to an R object}
}
\value{
R object
}
\description{
Use the QuickJS C API to convert a JSON string to an R object
}
