% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSContext.R
\name{JSContext}
\alias{JSContext}
\title{JSContext object}
\usage{
JSContext
}
\value{
A JSContext object containing an initialised JavaScript
context for evaluating scripts/commands
}
\description{
An initialised context within which to evaluate Javascript
scripts or commands.
}
