% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{ldflags}
\alias{ldflags}
\title{ldflags}
\usage{
ldflags(as_character = FALSE)
}
\arguments{
\item{as_character}{Whether the result should be returned as a string}
}
\description{
Function for returning the flags needed for linking to the package
}
