% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qjs.R
\name{qjs_eval}
\alias{qjs_eval}
\title{qjs_eval}
\usage{
qjs_eval(eval_string)
}
\arguments{
\item{eval_string}{A single string of the expression to evaluate}
}
\value{
The result of the provided expression
}
\description{
Evaluate a single Javascript expression.
}
\examples{
# Return the sum of two numbers:
qjs_eval("1 + 2")

# Concatenate strings:
qjs_eval("'1' + '2'")

# Create lists from objects:
qjs_eval("var t = {'a' : 1, 'b' : 2}; t")

}
