% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qjs.R
\name{qjs_passthrough}
\alias{qjs_passthrough}
\title{qjs_passthrough}
\usage{
qjs_passthrough(args, json = TRUE)
}
\arguments{
\item{args}{Args to pass through}

\item{json}{Whether to return a JSON string}
}
\value{
The input argument unchanged
}
\description{
Test function to pass through arguments
}
