% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{quickjs_flags}
\alias{quickjs_flags}
\alias{cppflags}
\alias{ldflags}
\title{QuickJS-NG Compiler and Linker Flags}
\usage{
cppflags()

ldflags()
}
\value{
A character string of flags
}
\description{
Return the flags needed for compiling against the bundled
quickjs-ng headers and library
}
