% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quickjs_version}
\alias{quickjs_version}
\title{Get the version of the bundled QuickJS library}
\usage{
quickjs_version()
}
\value{
Character string of the version of the bundled QuickJS library
}
\description{
Get the version of the bundled QuickJS library
}
