% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickjsr-package.R
\docType{package}
\name{QuickJSR-package}
\alias{QuickJSR-package}
\alias{QuickJSR}
\title{The QuickJSR package.}
\description{
An interface to the QuickJS lightweight Javascript engine
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrjohns/QuickJSR}
  \item \url{https://bellard.org/quickjs/}
  \item Report bugs at \url{https://github.com/andrjohns/QuickJSR/issues}
}

}
\author{
\strong{Maintainer}: Andrew R. Johnson \email{andrew.johnson@arjohnsonau.com} (\href{https://orcid.org/0000-0001-7000-8065}{ORCID})

Other contributors:
\itemize{
  \item QuickJS Authors (QuickJS sources and headers) [copyright holder]
  \item QuickJS-NG Authors (QuickJS-NG sources and headers) [copyright holder]
}

}
