% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terser.R
\name{terser}
\alias{terser}
\title{Minify JavaScript code using Terser}
\usage{
terser(code, options = list())
}
\arguments{
\item{code}{A character string containing the JavaScript code to be minified.}

\item{options}{A list of options to pass to Terser for minification.
See the \href{https://terser.org/docs/api-reference/#minify-options}{Terser documentation}
for an up-to-date list of options and default values.}
}
\value{
A list containing the minified code and any warnings or errors.
}
\description{
This function minifies JavaScript code using the Terser library.
}
\examples{
#' # Example usage
js_code <- "function add(a, b) { return a + b; }"
minified <- terser(js_code, list(sourceMap = TRUE))
minified$code
minified$map

js_code <- "console.log(add(1 + 2, 3 + 4));"
options <- list(
  toplevel = TRUE,
  compress = list(
    passes = 2,
    global_defs = list(
      "@console.log" = "alert"
    )
  ),
  format = list(
    preamble = "/* minified */"
  )
)
terser(js_code, options)$code
}
