% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qjs.R
\name{to_json}
\alias{to_json}
\title{to_json}
\usage{
to_json(arg)
}
\arguments{
\item{arg}{Argument to convert to JSON}
}
\value{
JSON string
}
\description{
Use the QuickJS C API to convert an R object to a JSON string
}
