% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{cxxflags}
\alias{cxxflags}
\title{cxxflags}
\usage{
cxxflags(to_console = FALSE)
}
\arguments{
\item{to_console}{Whether the result should be returned as a string}
}
\value{
Character string of CXX flags, or print flags to console
and invisibly return NULL (for use in package Makevars or similar)
}
\description{
Function for returning the C/C++ flags needed for compilation
using the package's headers
}
