% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSContext.R
\docType{data}
\name{JSContext-method-assign}
\alias{JSContext-method-assign}
\alias{assign}
\title{Assign a value to a variable in the current context}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
assign(var_name, value)
}
\arguments{
\item{var_name}{The name of the variable to assign}

\item{value}{The value to assign to the variable}
}
\value{
No return value, called for side effects
}
\description{
Assign a value to a variable in the current context
}
\examples{
\dontrun{
ctx <- JSContext$new()
ctx$assign("a", 1)
ctx$get("a")
}
}
\keyword{datasets}
