% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSContext.R
\docType{data}
\name{JSContext-method-call}
\alias{JSContext-method-call}
\alias{call}
\title{Call a JS function in the current context}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
call(function_name, ...)
}
\arguments{
\item{function_name}{The function to be called}

\item{...}{The arguments to be passed to the function}
}
\value{
The result of calling the specified function
}
\description{
Call a specified function in the JavaScript context with the
provided arguments.
}
\examples{
\dontrun{
ctx <- JSContext$new()
ctx$source(code = "function add(a, b) { return a + b; }")
ctx$call("add", 1, 2)
}
}
\keyword{datasets}
