% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSContext.R
\docType{data}
\name{JSContext-method-get}
\alias{JSContext-method-get}
\alias{get}
\title{Get a variable from the current context}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
get(var_name)
}
\arguments{
\item{var_name}{The name of the variable to retrieve}
}
\value{
The value of the variable
}
\description{
Get the value of a variable from the current context
}
\examples{
\dontrun{
ctx <- JSContext$new()
ctx$source(code = "var a = 1;")
ctx$get("a")
}
}
\keyword{datasets}
