% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSContext.R
\docType{data}
\name{JSContext-method-validate}
\alias{JSContext-method-validate}
\alias{validate}
\title{Assess validity of JS code without evaluating}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
validate(code_string)
}
\arguments{
\item{code_string}{The JS code to check}
}
\value{
A boolean indicating whether code is valid
}
\description{
Checks whether JS code string is valid code in the current context
}
\examples{
\dontrun{
ctx <- JSContext$new()
ctx$validate("1 + 2")
}
}
\keyword{datasets}
