% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSContext.R
\docType{data}
\name{JSContext}
\alias{JSContext}
\title{JSContext object}
\format{
An object of class \code{list} of length 1.
}
\usage{
JSContext
}
\value{
A JSContext object containing an initialised JavaScript
context for evaluating scripts/commands
}
\description{
An initialised context within which to evaluate Javascript
scripts or commands.
}
\keyword{datasets}
