% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{ldflags}
\alias{ldflags}
\title{ldflags}
\usage{
ldflags(to_console = FALSE)
}
\arguments{
\item{to_console}{Whether the result should be returned as a string}
}
\value{
Character string of linker flags, or print flags to console
and invisibly return NULL (for use in package Makevars or similar)
}
\description{
Function for returning the flags needed for linking to the package
}
