% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qjs.R
\name{to_json}
\alias{to_json}
\title{to_json}
\usage{
to_json(arg, auto_unbox = FALSE)
}
\arguments{
\item{arg}{Argument to convert to JSON}

\item{auto_unbox}{Automatically unbox single element vectors}
}
\value{
JSON string
}
\description{
Use the QuickJS C API to convert an R object to a JSON string
}
